+++ test/endToEnd/util/effectsTableDisplayTestUtil.ts
